refactor image upload to support SUSE images (bsc#1081855)
Currently, Ironic image upload logic only works for hLinux because the
assumptions it made are not true for SLES. For example, it was looking
for a ironic-deploy-images venv package, which is not provided by SuSE.
Instead, SuSE provides openstack-ironic-image. The file structure is
also different. For example, for SLES, the kernel and ramdisk images are
installed in /srv/tftpboot/openstack-ironic-image/ dir. The file names
and compression are also different.

This patch refactors the image upload logic so that it will work for
both. Notices that we are also making assumptions about the SuSE files.
See comments in the code.

Change-Id: I376e384254175b57862941e22f013d69aba809ab