remove excess space after the comma separator (bsc#1081857)
The space after the comma separator in the transport_url is
being interpreted liberally. Meaning, it's being interpreted as
part of the username. This causes authentication failure because
all the username/password are used to create a connection pool.

For example, if we have a transport_url that looks like this.

"rabbit://foo:bar@node1:6027, foo:bar@node2:6027"

The second connection will fail because user " foo" does not
exist.

This patch removes the excess space after the comma separator.

Change-Id: Ida9c6c0ff6c751b51e47503e8d1e2008a5832a70