Fix permission for /tftpboot dir (bsc#1083211)
Looks like we do need the "Allow from all" per

https://httpd.apache.org/docs/2.4/mod/mod_access_compat.html#allow

Otherwise, we'll get HTTP 403 when accessing anything in /tftpboot
from the baremetal nodes.

NOTE: we are using the same dir as image path for both tftp and http boot (i.e. ipxe). When booting over http, the images in this dir must be accessible by the baremetal node.

Change-Id: I7638aad0a862b2845796a6b1ac8c3b09bc0ccfc5