enable agent and oneview driver configuration(SCRD-6972)
Since we are no longer using class driver configuration, we need to
make sure we can still support selective agent and oneview driver
enablement. This patch enables the relevancy of agent_drivers_enabled
and oneview_enabled config options. Note that in the new driver
composition paradigm, agent and oneview shared certain types of
interfaces with other drivers.

This patch also removed the [conductor]auth_url option as it is
deprecated.

Change-Id: I68a3c7c2665da4df743ad0ee564850e300ca4b95