increase the Swift tempURL duration (SCRD-8290)
Need to increase the Swift tempURL duration from 2400 seconds (default)
to 1 hour since the ProLiant baremetal nodes we are using for QE testing
are dog slow. Swift tempURL expired before deploy was able to finish.

We'll also need to documentation this configuration so that customers
can make the appropriate calibrations depending on their hardware
performance.

Change-Id: I23b4cace08ca2615e6d901a249691860504667e7