increase url_timeout when talking to neutron (SCRD-8512)
Ironic multi-tenancy involves creating the port in the switch
and successfully bind to that port. This is all done as part of the
port creation API in Neutron. If the iLO network is slow or the
switch is slow to respond to SSH requests, the Neutron API will timed
out before the port creation is finished. The default URL timeout is
30 seconds when talking to Neutron APIs. In some cases, it is too short
as we've seen that in the lab the HPE compware switch took over 30
seconds to create/bind the port. Therefore, we need to bump it up
to 60 seconds. This should be "good enough" in most cases. However,
we'll also need to document the steps for customers to tune this
property if needed.

Change-Id: I71ec7f166de2e92171fb3f1c28fc5ff849c75536